[Refactor][Workloads] Align workloads/ops/ naming and structure with tileops/ops/
Drop _fwd suffix from mamba workloads (da_cumsum, ssd_chunk_scan,
ssd_chunk_state, ssd_state_passing), attention workloads (deepseek_nsa,
deepseek_nsa_cmp, gqa_sliding_window, gqa_sliding_window_varlen),
and deltanet/gated_deltanet workloads.

Rename mean_pooling_ops.py to mean_pooling.py.

Move attention workloads into workloads/ops/attention/ subpackage to
mirror the tileops/ops/attention/ layout.

Update all imports in tests/, benchmarks/, and docs/.
