+++ benchmarks/ops/attention/bench_deepseek_nsa.py
@@ -5,7 +5,7 @@ import torch
 
 from benchmarks.benchmark import BenchmarkBase, BenchmarkReport
 from tileops.ops import NSAFwdVarlenOp
-from workloads.ops.deepseek_nsa_fwd import NsaFwdTest
+from workloads.ops.attention.deepseek_nsa import NsaFwdTest
 
 
 class _NsaFwdTestBaseline(NsaFwdTest):
