+++ benchmarks/ops/attention/bench_deepseek_nsa_cmp.py
@@ -6,7 +6,7 @@ import torch
 from benchmarks.benchmark import BenchmarkBase, BenchmarkReport
 from tileops.ops import NSACmpFwdVarlenOp
 from workloads.nsa_utils import prepare_chunk_offsets
-from workloads.ops.deepseek_nsa_cmp_fwd import NsaCmpFwdTest
+from workloads.ops.attention.deepseek_nsa_cmp import NsaCmpFwdTest
 
 
 def _parallel_nsa_compression_fwd_pytorch(test, q, k_cmp, v_cmp, block_size, scale, offsets):
