+++ benchmarks/ops/attention/bench_mean_pooling.py
@@ -6,7 +6,7 @@ import torch
 from benchmarks.benchmark import BenchmarkBase, BenchmarkReport
 from tileops.ops import MeanPoolingForwardOp
 from workloads.nsa_utils import prepare_chunk_indices
-from workloads.ops.mean_pooling_ops import MeanPoolingTest
+from workloads.ops.attention.mean_pooling import MeanPoolingTest
 
 
 class _MeanPoolingTestBaseline(MeanPoolingTest):
