+++ tests/ops/attention/test_gqa_decode_paged.py
@@ -10,7 +10,9 @@ from torch.nn.attention import SDPBackend, sdpa_kernel
 
 from tests.test_base import FixtureBase, TestBase
 from tileops.ops import GroupedQueryAttentionDecodePagedWithKVCacheFwdOp
-from workloads.ops.gqa_decode_paged import GqaDecodePagedTest as _GqaDecodePagedTestWorkload
+from workloads.ops.attention.gqa_decode_paged import (
+    GqaDecodePagedTest as _GqaDecodePagedTestWorkload,
+)
 
 
 class GqaDecodePagedTest(_GqaDecodePagedTestWorkload, TestBase):
