+++ tests/ops/attention/test_mean_pooling.py
@@ -6,7 +6,7 @@ import torch
 from tests.test_base import FixtureBase, TestBase
 from tileops.ops import MeanPoolingForwardOp
 from workloads.nsa_utils import prepare_chunk_indices
-from workloads.ops.mean_pooling_ops import MeanPoolingTest as _MeanPoolingTestWorkload
+from workloads.ops.attention.mean_pooling import MeanPoolingTest as _MeanPoolingTestWorkload
 
 
 class MeanPoolingTest(_MeanPoolingTestWorkload, TestBase):
