+++ workloads/ops/da_cumsum.py
