+++ workloads/ops/ssd_state_passing.py
